feat(search): replace legacy Search by SearchRework and remove use of the homepage rework ff
